Add historical date ranges, dummy out push
Only added to the HTTP interface. Considering if there's an easy way
to have both the cmd line and the web interface driven from a simple
common object rather than duplicating stuff.

Initially mocking out the actual push for now so the system tests run
more quickly. Would like to do this properly but think it's easier to
wait until we have yaml file scenario support and just use that.